expand to 2,200,000,000, add unit tests, refine error msg

+++ src/main/java/com/chase/app/roman/service/IRomanNumeralService.java
@@ -1,5 +1,5 @@
 package com.chase.app.roman.service;
 
 public interface IRomanNumeralService {
-    String calculateRomanNumeralFromInt(int roman);
+    String calculateRomanNumeralFromInt(long roman);
 }
